pyproject.toml
[tool.poetry]
name = "kafka-websockets"
version = "0.1.0"
description = ""
authors = ["Christophoros <bekos.christopher@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
aiokafka = "^0.7.2"
python-dotenv = "^0.20.0"
python-socketio = "^5.7.1"
uvicorn = {extras = ["standard"], version = "^0.18.2"}

[tool.poetry.dev-dependencies]
python-socketio = {extras = ["client"], version = "^5.7.1"}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
